Patches

+++ src/main/java/me/stivendarsi/orbit/command/CommandHandler.java
@@ -14,7 +14,7 @@ public class CommandHandler {
         manager.registerEventHandler(LifecycleEvents.COMMANDS, (event) -> {
             Commands commands = event.registrar();
             commands.register(Commands.literal("orbit").requires(source -> source.getSender().hasPermission("orbit.admin"))
-                            .then(Commands.argument("experience", integer(0)).executes(OrbitCommands::open))
+                            .then(Commands.argument("experience", integer(1)).executes(OrbitCommands::open))
                     .build());
         });
